[refactor] combine types where appropriate
We don't need separate types files when the structs are only access from
one place
