+++ internal/announcers/kafka.go
@@ -14,6 +14,24 @@ type Kafka struct {
 	In chan []byte
 }
 
+type Announcer interface {
+	Status(e *Status)
+	Stop()
+}
+
+// Status is the message sent to the payload tracker
+type Status struct {
+	Service     string    `json:"service"`
+	Source      string    `json:"source,omitempty"`
+	Account     string    `json:"account"`
+	RequestID   string    `json:"request_id"`
+	InventoryID string    `json:"inventory_id,omitempty"`
+	SystemID    string    `json:"system_id,omitempty"`
+	Status      string    `json:"status"`
+	StatusMsg   string    `json:"status_msg"`
+	Date        time.Time `json:"date"`
+}
+
 // NewStatusAnnouncer creates a new announcer and starts the producer
 func NewStatusAnnouncer(cfg *queue.ProducerConfig) *Kafka {
 	k := &Kafka{
