+++ internal/logger/logger.go
@@ -19,6 +19,16 @@ import (
 var Log *logrus.Logger
 var logLevel logrus.Level
 
+// CustomCloudwatch adds hostname and app name
+type CustomCloudwatch struct {
+	Hostname string
+}
+
+//Marshaler is an interface any type can implement to change its output in our production logs.
+type Marshaler interface {
+	MarshalLog() map[string]interface{}
+}
+
 // NewCloudwatchFormatter creates a new log formatter
 func NewCloudwatchFormatter() *CustomCloudwatch {
 	f := &CustomCloudwatch{}
