+++ internal/stage/s3/s3.go
@@ -18,6 +18,14 @@ var (
 	client   *s3.S3
 )
 
+// Stager provides the mechanism to stage a payload via aws S3
+type Stager struct {
+	Bucket   string
+	Sess     *session.Session
+	Uploader *s3manager.Uploader
+	Client   *s3.S3
+}
+
 func init() {
 	sess = session.Must(session.NewSession())
 	uploader = s3manager.NewUploader(sess)
